Handle unconventional millis/micros
We found that parquet files created from pyarrow have "unconventional" format for millis and micros. This change is not original it's a merge of 2 already proposed  PRs
https://github.com/ZJONSSON/parquetjs/pull/65
https://github.com/ZJONSSON/parquetjs/pull/45

We are struggling with those issues, it this something you will consider to merge? we can create our own fork but we believe is better having one library.